requirements.txt
openai
openai-whisper
torch
numpy==1.26.4
fastapi
uvicorn
# speech_recognition